Move checkDependencies into a separate file that doesn't require dependencies
