+++ src/index.js
@@ -24,9 +24,6 @@ main(argv);
 
 // Run
 async function main(argv) {
-  // Check if dependencies are installed
-  await checkDependencies();
-
   // Find index of `doc-detective` or `run` in argv
   const index = argv.findIndex(
     (arg) => arg.endsWith("doc-detective") || arg.endsWith("index.js")
@@ -79,40 +76,4 @@ async function main(argv) {
 
   // Output results
   await outputResults(config, output, results, { command });
-}
-
-async function checkDependencies() {
-  if (!fs.existsSync(path.resolve(process.cwd(), "node_modules"))) {
-    const ask = `
-  It looks like you haven't installed the local dependencies yet. Would you like to install them now? (yes/no): `;
-    const answer = prompt(ask);
-    if (answer.toLowerCase() === "yes") {
-      console.log("Installing dependencies. This may take a few minutes.");
-      const { spawn } = require("child_process");
-      const install = spawn("npm", ["install"], { stdio: "inherit" });
-      await new Promise((resolve, reject) => {
-        install.on("close", (code) => {
-          if (code !== 0) {
-            reject(
-              new Error(
-                `Failed to install dependencies (exit code${code}).` +
-                  `Try running 'npm install' manually and check for errors.`
-              )
-            );
-          } else {
-            resolve();
-          }
-        });
-        install.on("error", (error) => {
-          reject(new Error(`Failed to start 'npm install': ${error.message}`));
-        });
-      });
-      console.log("Dependencies installed successfully.");
-    } else {
-      console.log(
-        "Please install the dependencies manually using 'npm install'."
-      );
-      process.exit(1);
-    }
-  }
 }
